backend_ops/local-api-obsidian_vault/integrations/tool-box/flyde-docs-scraper/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Web scraping
scrapfly-sdk==0.1.0
scrapy==2.11.0
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2

# MCP and context management
mcp==0.1.0
context7==0.1.0

# Monitoring and logging
sentry-sdk==1.38.0
structlog==23.2.0

# Data processing
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0

# Database and storage
sqlalchemy==2.0.23
alembic==1.13.0
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0